+++ src/Z38/SwissPayment/TransactionInformation/SEPACreditTransfer.php
@@ -20,22 +20,26 @@ class SEPACreditTransfer extends CreditTransfer
     protected $creditorIBAN;
 
     /**
-     * @var BIC
+     * @var BIC|null
      */
     protected $creditorAgentBIC;
 
     /**
      * {@inheritdoc}
      *
-     * @param IBAN $creditorIBAN     IBAN of the creditor
-     * @param BIC  $creditorAgentBIC BIC of the creditor's financial institution
+     * @param IBAN     $creditorIBAN     IBAN of the creditor
+     * @param BIC|null $creditorAgentBIC BIC of the creditor's financial institution
      */
-    public function __construct($instructionId, $endToEndId, Money\EUR $amount, $creditorName, PostalAddressInterface $creditorAddress, IBAN $creditorIBAN, BIC $creditorAgentBIC)
+    public function __construct($instructionId, $endToEndId, Money\EUR $amount, $creditorName, PostalAddressInterface $creditorAddress, IBAN $creditorIBAN, BIC $creditorAgentBIC = null)
     {
         parent::__construct($instructionId, $endToEndId, $amount, $creditorName, $creditorAddress);
 
         $this->creditorIBAN = $creditorIBAN;
-        $this->creditorAgentBIC = $creditorAgentBIC;
+
+        if ($creditorAgentBIC !== null) {
+            @trigger_error('Setting the creditor agent BIC of SEPA payments is deprecated. The execution of the payment will be based on the IBAN.', E_USER_DEPRECATED);
+            $this->creditorAgentBIC = $creditorAgentBIC;
+        }
     }
 
     /**
@@ -45,10 +49,11 @@ class SEPACreditTransfer extends CreditTransfer
     {
         $root = $this->buildHeader($doc, $paymentInformation, null, 'SEPA');
 
-        $creditorAgent = $doc->createElement('CdtrAgt');
-
-        $creditorAgent->appendChild($this->creditorAgentBIC->asDom($doc));
-        $root->appendChild($creditorAgent);
+        if ($this->creditorAgentBIC !== null) {
+            $creditorAgent = $doc->createElement('CdtrAgt');
+            $creditorAgent->appendChild($this->creditorAgentBIC->asDom($doc));
+            $root->appendChild($creditorAgent);
+        }
 
         $root->appendChild($this->buildCreditor($doc));