+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Allow transfers in Euro for payment types 2 and 3.
   * Deprecate `Z38\SwissPayment\BC` in favor of `Z38\SwissPayment\IID`.
   * Added support for ISR payments (type 1).
+  * Deprecate setting the creditor agent BIC of SEPA payments.
 
 ## 0.3.0 (2016-01-01)